Initial refactor. #84
