+++ src/main/java/net/derfla/quickeconomy/Main.java
@@ -3,10 +3,11 @@ package net.derfla.quickeconomy;
 import net.derfla.quickeconomy.command.BalanceCommand;
 import net.derfla.quickeconomy.command.BankCommand;
 import net.derfla.quickeconomy.command.QuickeconomyCommand;
+import net.derfla.quickeconomy.database.TableManagement;
+import net.derfla.quickeconomy.database.Utility;
 import net.derfla.quickeconomy.file.BalanceFile;
 import net.derfla.quickeconomy.listener.*;
 import net.derfla.quickeconomy.util.AccountCache;
-import net.derfla.quickeconomy.util.DatabaseManager;
 import net.derfla.quickeconomy.util.DerflaAPI;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -83,9 +84,9 @@ public class Main extends JavaPlugin {
     private void setupSQLMode() {
         getLogger().info("Running in SQL mode. Attempting to connect to SQL server...");
         try {
-            DatabaseManager.connectToDatabase();
+            Utility.connectToDatabase();
             SQLMode = true;
-            DatabaseManager.createTables();
+            TableManagement.createTables();
         } catch (Exception e) {
             getLogger().severe("Could not establish a database connection: " + e.getMessage());
             getServer().getPluginManager().disablePlugin(this);
@@ -95,8 +96,8 @@ public class Main extends JavaPlugin {
     @Override
     public void onDisable() {
         // Plugin shutdown logic
-        DatabaseManager.closePool();
-        DatabaseManager.shutdownExecutorService(); // Shutdown async thread handler (for database operations)
+        Utility.closePool();
+        Utility.shutdownExecutorService(); // Shutdown async thread handler (for database operations)
     }
 
     public static Main getInstance() {
