+++ src/main/java/net/derfla/quickeconomy/listener/PlayerJoinListener.java
@@ -1,6 +1,8 @@
 package net.derfla.quickeconomy.listener;
 
 import net.derfla.quickeconomy.Main;
+import net.derfla.quickeconomy.database.Shop;
+import net.derfla.quickeconomy.database.TableManagement;
 import net.derfla.quickeconomy.util.*;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.event.ClickEvent;
@@ -56,7 +58,7 @@ public class PlayerJoinListener implements Listener {
 
                 // Handle empty shops list if enabled
                 if (Main.SQLMode && plugin.getConfig().getBoolean("shop.emptyShopListJoin")) {
-                    DatabaseManager.displayEmptyShopsView(uuid)
+                    Shop.displayEmptyShopsView(uuid)
                         .thenAccept(emptyShops -> {
                             if (emptyShops != null && !emptyShops.isEmpty()) {
                                 int emptyShopCount = emptyShops.size();
